Backend/requirements.txt
alembic==1.14.1
amqp==5.3.1
aniso8601==10.0.0
annotated-types==0.7.0
bcrypt==4.3.0
billiard==4.2.1
blinker==1.8.2
cached-property==2.0.1
cachelib==0.13.0
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
distlib==0.3.9
dnspython==2.6.1
dotenv==0.9.9
email_validator==2.2.0
filelock==3.16.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.3
Mako==1.3.9
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-enum==1.5.1
mpesa==0.0.1
msgspec==0.19.0
mypy_extensions==1.1.0
packaging==25.0
pipenv==2024.4.1
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.2
redis==6.0.0
requests==2.32.3
setuptools==80.4.0
six==1.17.0
slugify==0.0.1
SQLAlchemy==2.0.40
sqlalchemy-orm==1.2.10
SQLAlchemy-serializer==1.4.12
tenacity==9.1.2
text-unidecode==1.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
virtualenv==20.29.3
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.20.2
